requirements.txt
django==4.2.6
django-ckeditor==6.7.0
Pillow==10.1.0
djangorestframework==3.14.0
requests==2.31.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
# pip install -r requirements.txt
# pip list